feat: add tag lifecycle, cognitive drift detection, custom embeddings
Tag Lifecycle System:
- Confidence decay over time (high→medium→low→stale)
- Contradiction detection ("it's not there", "already checked")
- Relocation intent parsing ("I moved my pills to the bedroom")
- Item removal detection ("I finished my vitamins")
- Stale tag re-confirmation prompts

Cognitive Drift Detector:
- Repeat question detection via cosine similarity
- Confusion signal matching (trailing off, "what was I looking for")
- Gentle redirect generation for disoriented patients
- Caregiver alert triggers (4+ repeats/hr, distress signals)

Custom Embeddings (from scratch, no ML libraries):
- TF-IDF with hand-computed term frequency and IDF
- BM25 (Okapi) with full formula implementation
- L2 normalization, cosine similarity, euclidean distance
- Domain-specific synonym expansion for vocabulary bridging
- Hybrid retriever with weighted TF-IDF + BM25 combination
- Suffix-stripping stemmer with improved rules

Also: float coordinates for Konva.js, confidence CHECK constraint,
caregiver alerts router, updated query flow with full lifecycle

+++ backend/routers/floorplan.py
@@ -46,8 +46,8 @@ async def suggest_tag_positions(
             TagPlacement(
                 tag_id=p["tag_id"],
                 label=p["label"],
-                x=max(0, min(int(p["x"]), request.room_width_px)),
-                y=max(0, min(int(p["y"]), request.room_height_px)),
+                x=max(0.0, min(float(p["x"]), float(request.room_width_px))),
+                y=max(0.0, min(float(p["y"]), float(request.room_height_px))),
             )
         )
